Bug 1062657 - Tweak SourceBuffer::AppendData overlap algorithm and switch internal SourceBuffer timestamp handling from double/seconds to int64_t/microseconds.  r=karlt
UltraBlame original commit: f63cd36acc652bcff8072c708e521358bfc58f63

+++ content/media/mediasource/TrackBuffer.h
@@ -64,9 +64,9 @@ public:
   bool IsReady();
 
   
-  void LastTimestamp(double& aStart, double& aEnd);
-  void SetLastStartTimestamp(double aStart);
-  void SetLastEndTimestamp(double aEnd);
+  void LastTimestamp(int64_t& aStart, int64_t& aEnd);
+  void SetLastStartTimestamp(int64_t aStart);
+  void SetLastEndTimestamp(int64_t aEnd);
 
   
   
@@ -122,8 +122,8 @@ private:
 
   
   
-  double mLastStartTimestamp;
-  double mLastEndTimestamp;
+  int64_t mLastStartTimestamp;
+  int64_t mLastEndTimestamp;
